railway/scripts/services/requirements.txt
uvicorn[standard]
fastapi[all]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp